Adding gpu_conv_algorithm_picker base class
