+++ tensorflow/compiler/xla/service/gpu/BUILD
@@ -596,10 +596,14 @@ cc_library(
 
 cc_library(
     name = "gpu_conv_algorithm_picker",
-    srcs = []
-    + if_rocm_is_configured(["miopen_conv_algorithm_picker.cc",])
-    + if_cuda_is_configured(["cudnn_conv_algorithm_picker.cc",]),
-    hdrs = []
+    srcs = ["gpu_conv_algorithm_picker.cc"]
+    + if_rocm_is_configured([
+        "miopen_conv_algorithm_picker.cc",
+    ])
+    + if_cuda_is_configured([
+        "cudnn_conv_algorithm_picker.cc",
+    ]),
+    hdrs = ["gpu_conv_algorithm_picker.h"]
     + if_rocm_is_configured(["miopen_conv_algorithm_picker.h",])
     + if_cuda_is_configured(["cudnn_conv_algorithm_picker.h",]),
     deps = [
